+++ packages/niivue/src/drawing/index.ts
@@ -1,2 +1,13 @@
-export { drawUndo } from './undo.js'
-export { encodeRLE, decodeRLE } from './rle.js'
+export { drawUndo } from '@/drawing/undo'
+export { encodeRLE, decodeRLE } from '@/drawing/rle'
+export {
+  findBoundarySlices,
+  extractSlice,
+  extractIntensitySlice,
+  insertColorMask,
+  doGeometricInterpolation,
+  doIntensityGuidedInterpolation,
+  smoothSlice,
+  calculateIntensityWeight,
+  interpolateMaskSlices
+} from '@/drawing/masks'
