+++ packages/niivue/src/utils/index.ts
@@ -1,1 +1,5 @@
-export { nice } from './nice.js'
+export { nice } from '@/utils/nice'
+export { readFileAsDataURL } from '@/utils/file-utils'
+export { img2ras16, unpackFloatFromVec4i, intensityRaw2Scaled } from '@/utils/image-utils'
+export { tickSpacing, deg2rad, negMinMax, clamp } from '@/utils/math-utils'
+export { swizzleVec3, isRadiological, unProject } from '@/utils/webgl-utils'
